added 'me' variables
- added local 'me' variables which are used to alias names relative to
the current client (board names can now use the $me variable to
designate a layout that is relative to the current client)
- split BaseProxy into BaseClientProxy and BaseServerProxy
- the BaseClientProxy now contains a board, as this is shared amongst
all clients that use that proxy
- multiple clients use the same proxy (i.e. the proxy represents all the
clients for that screen). this reduces the amount of data transfered if
there are multiple clients on the same screen (e.g. with a shared board)
- configs are now based on screens, with each screen operating in a
particular mode ('shared', 'pass-n-play' or 'networked'). Screens are
picked via a URI and permit a number of players to use the same screen
(for shared and pass-n-play modes).
- fixed gulp file to wait until compilation finished before merging with
gamex.js (renamed boardx.js)
- index.js updated to support new interfaces
- mancala game now calls each side red and blue, rather than p1 and p2
(as players can change)